pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archminer"
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "GPL-3.0-or-later"
keywords = []
authors = [
  { name = "Ben Companjen", email = "ben@companjen.name" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "pdfminer.six>=20220524"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/unknown/archminer#readme"
Issues = "https://github.com/unknown/archminer/issues"
Source = "https://github.com/unknown/archminer"

[project.scripts]
archminer = "archminer.cli:cli"

[tool.hatch.version]
path = "archminer/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=archminer --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "archminer/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
